Go back to single virtualenv for unit testing
