+++ tests/unit/records/schema/field/test_field.py
@@ -45,7 +45,7 @@ class TestField(unittest.TestCase):
                                processing_instructions=mock_processing_instructions)
         self.assertEqual(out, mock_field_from_index.return_value)
 
-    @patch('records_mover.records.schema.field.field_from_sqlalchemy_column')
+    @patch('records_mover.records.schema.field.sqlalchemy.field_from_sqlalchemy_column')
     def test_from_sqlalchemy_column(self, mock_field_from_sqlalchemy_column):
         mock_column = Mock(name='column')
         mock_driver = Mock(name='driver')
@@ -59,7 +59,7 @@ class TestField(unittest.TestCase):
                                rep_type=mock_rep_type)
         self.assertEqual(out, mock_field_from_sqlalchemy_column.return_value)
 
-    @patch('records_mover.records.schema.field.field_to_sqlalchemy_type')
+    @patch('records_mover.records.schema.field.sqlalchemy.field_to_sqlalchemy_type')
     def test_to_sqlalchemy_type(self, mock_field_to_sqlalchemy_type):
         mock_driver = Mock(name='driver')
         mock_name = Mock(name='name')
@@ -76,7 +76,7 @@ class TestField(unittest.TestCase):
         mock_field_to_sqlalchemy_type.assert_called_with(field, mock_driver)
         self.assertEqual(out, mock_field_to_sqlalchemy_type.return_value)
 
-    @patch('records_mover.records.schema.field.field_to_sqlalchemy_column')
+    @patch('records_mover.records.schema.field.sqlalchemy.field_to_sqlalchemy_column')
     def test_to_sqlalchemy_column(self, mock_field_to_sqlalchemy_column):
         mock_driver = Mock(name='driver')
         mock_name = Mock(name='name')
