+++ tests/unit/url/test_s3_directory.py
