--- tests/unitalldeps/test_public_interface.py
@@ -1,1 +1,0 @@
-from records_mover.airflow.hooks import records_hook  # noqa
